scripts/admin/requirements.txt
flask>=2.0
pyyaml>=6.0